refactor(api): move usecases
to src/team
