+++ api/src/team/domain/usecases/create-membership.usecase.js
@@ -1,5 +1,5 @@
-import { OrganizationArchivedError } from '../errors.js';
-import { roles } from '../models/Membership.js';
+import { OrganizationArchivedError } from '../../../../lib/domain/errors.js';
+import { roles } from '../../../../lib/domain/models/Membership.js';
 
 const createMembership = async function ({ userId, organizationId, membershipRepository, organizationRepository }) {
   const organization = await organizationRepository.get(organizationId);
