+++ api/tests/team/unit/domain/usecases/create-membership.usecase.test.js
@@ -1,9 +1,9 @@
-import { OrganizationArchivedError } from '../../../../lib/domain/errors.js';
-import { Membership } from '../../../../lib/domain/models/Membership.js';
-import { createMembership } from '../../../../lib/domain/usecases/create-membership.js';
-import { catchErr, domainBuilder, expect, sinon } from '../../../test-helper.js';
+import { OrganizationArchivedError } from '../../../../../lib/domain/errors.js';
+import { Membership } from '../../../../../lib/domain/models/Membership.js';
+import { createMembership } from '../../../../../src/team/domain/usecases/create-membership.usecase.js';
+import { catchErr, domainBuilder, expect, sinon } from '../../../../test-helper.js';
 
-describe('Unit | UseCase | create-membership', function () {
+describe('Unit | Team | Domain | UseCase | create-membership', function () {
   it('should insert a new membership with role ADMIN', async function () {
     // given
     const organization = domainBuilder.buildOrganization();
